Add correct reshaping for yuv422p frames
The raw buffer coming from ffmpeg was assumed to be ordered in a
HEIGHTxWIDTHxCOLOR way.
This is not the case for YUV frames which are ordered in a
COLORxHEIGHTxWIDTH way.
The _readFrame function from VideoReaderAbstract has been patched
for yuv444p raw frames.

YUV formats with chroma subsampling do not have the same amount of
pixel per each channel.
A warning message has been added for users using other pixel formats
than yuv444p and rgb24.